+++ pkg/command/echo.go
