+++ pkg/command/echo_test.go
