+++ pkg/command/help_test.go
