+++ pkg/command/image.go
