+++ pkg/command/score.go
