+++ pkg/command/select_test.go
