+++ pkg/command/service.go
