+++ pkg/command/service_test.go
